# planning/api-documentations/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "echomemory backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/admin/list": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Admin List Users",
        "description": "以管理员身份列出用户，支持筛选。",
        "operationId": "admin_list_users_api_v1_users_admin_list_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 3,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 3,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按用户名或昵称搜索",
              "title": "Q"
            },
            "description": "按用户名或昵称搜索"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserMeOut"
                  },
                  "title": "Response Admin List Users Api V1 Users Admin List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/playlists/": {
      "post": {
        "tags": [
          "playlists"
        ],
        "summary": "Create Playlist",
        "description": "创建歌单。\n\n接收 multipart/form-data，可选上传封面图片。\n标签由系统根据歌曲收藏自动计算，不允许手动编辑。",
        "operationId": "create_playlist_api_v1_playlists__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_playlist_api_v1_playlists__post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "playlists"
        ],
        "summary": "List Playlists",
        "description": "查询我的歌单列表（按创建时间倒序）。",
        "operationId": "list_playlists_api_v1_playlists__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistListOut"
                  },
                  "title": "Response List Playlists Api V1 Playlists  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/playlists/{playlist_id}": {
      "get": {
        "tags": [
          "playlists"
        ],
        "summary": "Get Playlist",
        "description": "获取歌单详情。\n\n仅允许查看自己的歌单或 `is_private=False` 的公开歌单。",
        "operationId": "get_playlist_api_v1_playlists__playlist_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Playlist Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "playlists"
        ],
        "summary": "Update Playlist",
        "description": "修改歌单信息（仅文本字段，不含封面替换和标签编辑）。",
        "operationId": "update_playlist_api_v1_playlists__playlist_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Playlist Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaylistUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "playlists"
        ],
        "summary": "Delete Playlist",
        "description": "删除自己的歌单。",
        "operationId": "delete_playlist_api_v1_playlists__playlist_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Playlist Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/playlists/{playlist_id}/musics/{music_id}": {
      "post": {
        "tags": [
          "playlists"
        ],
        "summary": "Add Music To Playlist",
        "description": "添加一首已上架音乐到歌单。仅允许操作自己的歌单。",
        "operationId": "add_music_to_playlist_api_v1_playlists__playlist_id__musics__music_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Playlist Id"
            }
          },
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "playlists"
        ],
        "summary": "Remove Music From Playlist",
        "description": "从歌单移除一首音乐。仅允许操作自己的歌单。",
        "operationId": "remove_music_from_playlist_api_v1_playlists__playlist_id__musics__music_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Playlist Id"
            }
          },
          {
            "name": "music_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Music Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/collections/playlists/{playlist_id}": {
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Collect Playlist",
        "description": "收藏歌单。已收藏则静默返回。",
        "operationId": "collect_playlist_api_v1_collections_playlists__playlist_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Playlist Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistCollectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "collections"
        ],
        "summary": "Uncollect Playlist",
        "description": "取消收藏歌单。未收藏则静默成功。",
        "operationId": "uncollect_playlist_api_v1_collections_playlists__playlist_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Playlist Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/playlists": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "List Playlist Collections",
        "description": "查询我的收藏歌单列表（按收藏时间倒序）。",
        "operationId": "list_playlist_collections_api_v1_collections_playlists_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistCollectionOut"
                  },
                  "title": "Response List Playlist Collections Api V1 Collections Playlists Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_playlist_api_v1_playlists__post": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private",
            "default": false
          },
          "cover_icon": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Icon"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "Body_create_playlist_api_v1_playlists__post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlaylistListOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private"
          },
          "cover_icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Icon Url"
          },
          "user": {
            "$ref": "#/components/schemas/PlaylistUserOut"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "is_private",
          "user",
          "created_at"
        ],
        "title": "PlaylistListOut",
        "description": "歌单列表项输出（精简）。"
      },
...
      "PlaylistOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private"
          },
          "cover_icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Icon Url"
          },
          "collect_count": {
            "type": "integer",
            "title": "Collect Count"
          },
          "play_count": {
            "type": "integer",
            "title": "Play Count"
          },
          "hot": {
            "type": "integer",
            "title": "Hot"
          },
          "comment_count": {
            "type": "integer",
            "title": "Comment Count"
          },
          "is_like": {
            "type": "boolean",
            "title": "Is Like"
          },
          "is_recommended": {
            "type": "boolean",
            "title": "Is Recommended"
          },
          "user": {
            "$ref": "#/components/schemas/PlaylistUserOut"
          },
          "musics": {
            "items": {
              "$ref": "#/components/schemas/PlaylistMusicOut"
            },
            "type": "array",
            "title": "Musics",
            "default": []
          },
          "emotion_tags": {
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "type": "array",
            "title": "Emotion Tags",
            "default": []
          },
          "interest_tags": {
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "type": "array",
            "title": "Interest Tags",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "is_private",
          "collect_count",
          "play_count",
          "hot",
          "comment_count",
          "is_like",
          "is_recommended",
          "user",
          "created_at",
          "updated_at"
        ],
        "title": "PlaylistOut",
        "description": "歌单详情输出。"
      },
      "PlaylistUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_private": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Private"
          }
        },
        "type": "object",
        "title": "PlaylistUpdate",
        "description": "修改歌单信息的请求体（不含封面替换和标签编辑）。"
      },
...
      "UserMeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          },
          "gender": {
            "type": "integer",
            "title": "Gender"
          },
          "role": {
            "type": "integer",
            "title": "Role"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "exp": {
            "type": "integer",
            "title": "Exp"
          },
          "city_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Id"
          },
          "birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified"
          },
          "like_count": {
            "type": "integer",
            "title": "Like Count"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "safety_score": {
            "type": "integer",
            "title": "Safety Score"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "banned_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banned At"
          },
          "ban_duration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ban Duration"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "nickname",
          "gender",
          "role",
          "level",
          "exp",
          "is_verified",
          "like_count",
          "status",
          "safety_score"
        ],
        "title": "UserMeOut"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}